Add support for handling a group of controls.
These changes allow Happy.js to be happy when validating a collection of controls, e.g., $('[name="radios"]').

The current version of the code adds the error label against each control in the collection, but removes it only from the first child (`.get(0)` on line 74).

The second change allow fields to override the default validation event. Useful, again, for radios/checkboxes where you'd like to have `when: 'click'`, but still keep the default `blur` for the rest of the controls.